+++ package.json
@@ -18,6 +18,7 @@
     "@types/react-dom": "18.0.10",
     "eslint": "8.33.0",
     "eslint-config-next": "^13.2.4",
+    "hamburger-react": "^2.5.0",
     "next": "^13.2.4",
     "next-sanity": "^4.1.2",
     "react": "^18.2.0",
