+++ src/components/Hamburger.tsx
@@ -1,0 +1,6 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { CommonBurgerProps, Squash as SquashBurgerMenu } from 'hamburger-react';
+
+export const SquashHamburger = (props: CommonBurgerProps) => (
+  <SquashBurgerMenu {...props} />
+);
